requirements.txt
adal==1.2.7
applicationinsights==0.11.10
appnope==0.1.3
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azure-storage-blob==12.11.0
azure-storage-queue==12.2.0
azureml-automl-core==1.40.0
azureml-automl-runtime==1.40.0
azureml-core==1.40.0.post2
azureml-dataprep==3.0.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.4.3
azureml-dataset-runtime==1.40.0
azureml-defaults==1.40.0
azureml-inference-server-http==0.4.13
azureml-interpret==1.40.0
azureml-mlflow==1.40.0
azureml-telemetry==1.40.0
azureml-train-automl==1.40.0
azureml-train-automl-client==1.40.0
azureml-train-automl-runtime==1.40.0.post2
azureml-train-core==1.40.0
azureml-train-restclients-hyperdrive==1.40.0
azureml-training-tabular==1.40.0
azureml-widgets==1.40.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bokeh==2.4.2
boto==2.49.0
boto3==1.20.19
botocore==1.23.19
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.0.0
configparser==3.7.4
contextlib2==21.6.0
cryptography==36.0.2
Cython==0.29.17
dask==2.30.0
databricks-cli==0.16.6
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distributed==2.30.1
distro==1.7.0
docker==5.0.3
dotnetcore2==2.1.23
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fire==0.4.0
Flask==1.0.3
flatbuffers==2.0
fsspec==2022.3.0
fusepy==3.0.1
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.2
google-auth==2.6.6
googleapis-common-protos==1.56.0
gunicorn==20.1.0
HeapDict==1.0.1
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inference-schema==1.3.0
interpret-community==0.24.2
interpret-core==0.2.7
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==1.1.0
jedi==0.18.1
jeepney==0.8.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras2onnx==1.6.0
knack==0.9.0
lightgbm==3.2.1
llvmlite==0.36.0
locket==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
ml-wrappers==0.0.6
mlflow-skinny==1.25.1
msal==1.17.0
msal-extensions==0.3.1
msgpack==1.0.3
msrest==0.6.21
msrestazure==0.6.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
nimbusml==1.8.0
notebook==6.4.11
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.0
onnx==1.7.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.8.1
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
paramiko==2.10.3
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkginfo==1.8.2
pmdarima==1.7.1
portalocker==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==0.22.2.post1
scipy==1.5.2
SecretStorage==3.3.2
Send2Trash==1.8.0
shap==0.39.0
six==1.16.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==1.9.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.11.1
tabulate==0.8.9
tblib==1.7.0
termcolor==1.1.0
terminado==0.13.3
tinycss2==1.1.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.12.1
zict==2.1.0
zipp==3.8.0
